Test Car-hit function

+++ [HWBEM] Code/HWBEM_Project/Ultilities/inc/UART.h
@@ -13,7 +13,7 @@
 void UART_InitController();
 void UART_SendByte(uint8_t _data);
 void UART_SendString(uint8_t *_data);
-void UART_SendNumber(uint16_t _number);
+void UART_SendNumber(uint32_t _number);
 void UART_ISR(void);
 uint8_t UART_ReceiveByte();
 uint8_t UART_RXAvailable();
